lambda_layer_build/requirements.txt
pdfminer.six==20231228
chardet==5.2.0
cryptography==41.0.7
